added exercise page
created the page where the exercise preview page would go to, where the actual exercises will reside. Also made it so that the 'Start Exercise' button from the Exercise Preview page linked to this new page, and made sure that the you could view your feedback after doing the exercise.